Update on the main -> use of an example
Our main does not work with CoxRossRubinstein_2. The link is not done and we don't understand why. Because we cannot test our binomial tree, we cannot be sure how to change the j parameter in the underlying.
Stangely, the main works when we choose CoxRossRubinstein but then it uses the the binomialtree of QL installed on the PC.